chart dan bug fixing

+++ src/client/src/components/UserList.js
@@ -6,10 +6,31 @@ const UserList = ({ onEdit, refresh }) => {
     const [users, setUsers] = useState([]);
 
     useEffect(() => {
+        // console.log(); 
         // fetchUsers list in here
+        fetchUsers(); 
     }, [refresh]); // Fetch data when `refresh` prop changes
 
+    const fetchUsers = async () => {
+        // setLoading(true);
+        try {
+            const response = await fetch('/api/users');
+            if (!response.ok) throw new Error('Failed to fetch users');
+
+            const data = await response.json();
+            setUsers(data);
+        } catch (error) {
+            console.error('Error fetching users:', error);
+        } finally {
+            // setLoading(false);
+        }
+    };
+
     const handleDelete = (id) => {
+
+        const isConfirmed = window.confirm("Are you sure you want to delete this user?");
+        if (!isConfirmed) return;
+        
         $.ajax({
             url: `/api/users/${id}`,
             type: 'DELETE',
